Addressed Pull Requests comments.
